make single precision default

+++ HeatSolve.f95
@@ -6,19 +6,19 @@ program heat_solver
 	implicit none
 
 	!Declare Variables
-	real(8), parameter :: D = 7 * 10**(-7)				!Diffusivity (m2s-1) 
-	real(8), parameter :: t_start = 0					!Start time (s)
-	real(8), parameter :: t_stop = 60*365*24*60*60			!Stop time (s)
-	real(8), parameter :: H = 60					!Column height (m)
-	real(8), parameter :: dt = 1800 					!Time step (s)
-	real(8), parameter :: dz = 1					!Space step (m)
-	real(8), parameter :: xi = dt * D * dz**(-2)
+	real, parameter :: D = 7 * 10**(-7)				!Diffusivity (m2s-1) 
+	real, parameter :: t_start = 0					!Start time (s)
+	real, parameter :: t_stop = 60*365*24*60*60			!Stop time (s)
+	real, parameter :: H = 60					!Column height (m)
+	real, parameter :: dt = 1800 					!Time step (s)
+	real, parameter :: dz = 1					!Space step (m)
+	real, parameter :: xi = dt * D * dz**(-2)
 	
 	!Initialize array for temperature
 	integer, parameter :: m = H / dz + 1
 	integer, parameter :: n = (t_stop - t_start) / dt
 	integer :: i, j
-	real(8), dimension(m, n) :: T					!Temperature
+	real, dimension(m, n) :: T					!Temperature
 
 	!Set initial condition
 	do i = 1, m
